requirements.txt
torch
torchvision
transformers
streamlit
Pillow
gTTS
matplotlib
pandas
numpy
